post

+++ post/post.php
@@ -21,20 +21,29 @@
           <tbody class="w-100">
             <?php
             foreach ($post as $elemento) {
-              echo '<tr>
-              <td><h5>' .$elemento['title'].'</h5></td>
-              <td><h6>'.$elemento['brief'].'</h6></td>
-              <td><h6>'.$elemento['body'].'</h6></td>
-              <td><img src="../img/'.$elemento['image'].'" height="50px"  /></td>
-              <td class="text-right"><a href="eliminar.php?id='.$elemento['id'].'" class="btn btn-primary">Eliminar este post</a>
-              <a href="form.php?id='.$elemento['id'].'" class="btn btn-success">Editar, este post</a></td>
-              </tr>';}
-              ?>
+
+              if ((isset($_GET['lang']) && $_GET['lang'] ==$elemento['language_id']) || !isset($_GET['lang']) ){
+                echo '<tr>
+                <td><h5>' .$elemento['title'].'</h5></td>
+                <td><h6>'.$elemento['brief'].'</h6></td>
+                <td><h6>'.$elemento['body'].'</h6></td>
+                <td>';
+                if($elemento['image'] != ''){
+                  echo '<img src="../img/'.$elemento['image'].'" height="50px"  />';
+                }
+                echo '
+                </td>
+                <td class="text-right"><a href="eliminar.php?id='.$elemento['id'].'" class="btn btn-primary">Eliminar este post</a>
+                <a href="form.php?id='.$elemento['id'].'" class="btn btn-success">Editar, este post</a></td>
+                </tr>';
+              }
+            }
+            ?>
             </tbody>
           </table>
         </div>
         <div class="col-6">
-          <a href="post.php" class="btn btn-warning btn-lg">Agregar un post</a>
+          <a href="form.php" class="btn btn-warning btn-lg">Agregar un post</a>
 
         </div>
       </div>
